implement the new summary list card component and remove old table row filter logic
